[master]google规范格式化

+++ nas-common/src/main/java/com/minyan/nascommon/vo/MActivityChannelDetailVO.java
@@ -9,8 +9,8 @@ import lombok.Data;
  */
 @Data
 public class MActivityChannelDetailVO {
-    private Long activityId;
-    private String channelName;
-    private String channelCode;
-    private String channelUrl;
+  private Integer activityId;
+  private String channelName;
+  private String channelCode;
+  private String channelUrl;
 }
